requirements.txt
alembic
SQLAlchemy
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
pydantic==2.7.4
PyJWT==2.8.0
starlette==0.37.2
celery==5.4.0
beanie==1.26.0
httpx==0.27.0
loguru==0.7.2
pydantic-settings==2.3.3
git+https://github.com/aio-libs/aiocache.git@master